modifed fronted part of docker
