+++ docker/model_service/Dockerfile
@@ -12,9 +12,6 @@ COPY ./src/backend/toxic.csv /code/
 
 RUN poetry config virtualenvs.create false && poetry install --no-interaction --no-ansi
 
-RUN npm install
-RUN npm run build
-
 EXPOSE 8000
   
 # CMD ["uvicorn", "backend.main:app", "--host", "0.0.0.0", "--port", "8000"]
