--- docker/nginx.conf
@@ -1,34 +1,0 @@
-user  nginx;
-worker_processes  auto;
-
-error_log  /var/log/nginx/error.log warn;
-pid        /var/run/nginx.pid;
-
-events {
-    worker_connections  4096;
-}
-
-http {
-    include       /etc/nginx/mime.types;
-    default_type  application/octet-stream;
-
-    log_format  main  '$remote_addr - $remote_user [$time_local] "$request" '
-                      '$status $body_bytes_sent "$http_referer" '
-                      '"$http_user_agent" "$http_x_forwarded_for"';
-
-    access_log  /var/log/nginx/access.log  main;
-    sendfile        on;
-    keepalive_timeout  65;
-	
-	server { 
-	listen 9000;
-	server_name name.ru;
-	root /src/frontend/pages;
-
-	index index.php index.html index.html;
-
-	location / {
-	try_files $uri $uri/ /index.html;
-		}
-	}
-}
